- Deleted obsolete tests - Moved characterattributetype tests
